requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
banal==1.0.6
bert-score==0.3.13
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataset==1.6.2
datasets==3.2.0
decorator==5.1.1
dill==0.3.8
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
greenlet==3.1.1
grpcio==1.70.0
huggingface-hub==0.28.1
hydra-core==1.3.2
idna==3.10
inflate64==1.0.1
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
loralib==0.1.2
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==0.22.0
pyarrow==19.0.0
pybcj==1.0.3
pyclean==3.1.0
pycryptodomex==3.21.0
Pygments==2.19.1
pyparsing==3.2.1
pyppmd==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzstd==0.16.2
regex==2024.11.6
requests==2.32.3
rouge-score==0.1.2
safetensors==0.5.2
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
SQLAlchemy==1.4.54
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
texttable==1.7.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
